fix(ci): resolve critical compilation errors
- Fixed type redeclarations in pkg/llm
- Resolved type compatibility in pkg/nephio/porch
- Fixed undefined types in pkg/auth (JWT v5)
- Corrected lock copying errors in pkg/rag
- Removed unused variables and import cycles

+++ pkg/llm/cache.go
@@ -196,7 +196,7 @@ func (c *ResponseCache) Set(key, response string) {
 		LastAccess: time.Now(),
 		HitCount:   0,
 		TTL:        ttl,
-		Size:       len(response),
+		Size:       int64(len(response)),
 		Keywords:   keywords,
 		Level:      1,
 	}
